remove breadcrumbs object and command
