+++ settings/DevHome.Settings/ViewModels/PreferencesViewModel.cs
@@ -1,20 +1,22 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+using System.Collections.ObjectModel;
 using System.Threading.Tasks;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
-using DevHome.Common.Contracts;
 using DevHome.Common.Services;
-using DevHome.Common.TelemetryEvents;
 using DevHome.Contracts.Services;
-using DevHome.Telemetry;
+using DevHome.Settings.Models;
 using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
 
 namespace DevHome.Settings.ViewModels;
 
 public partial class PreferencesViewModel : ObservableObject
 {
+    public ObservableCollection<Breadcrumb> Breadcrumbs { get; }
+
     private readonly IThemeSelectorService _themeSelectorService;
 
     [ObservableProperty]
@@ -23,8 +25,24 @@ public partial class PreferencesViewModel : ObservableObject
     public PreferencesViewModel(IThemeSelectorService themeSelectorService)
     {
         _themeSelectorService = themeSelectorService;
-
         _elementTheme = _themeSelectorService.Theme;
+
+        var stringResource = new StringResource("DevHome.Settings/Resources");
+        Breadcrumbs = new ObservableCollection<Breadcrumb>
+        {
+            new(stringResource.GetLocalized("Settings_Header"), typeof(SettingsViewModel).FullName!),
+            new(stringResource.GetLocalized("Settings_Preferences_Header"), typeof(PreferencesViewModel).FullName!),
+        };
+    }
+
+    [RelayCommand]
+    public void BreadcrumbBarItemClicked(BreadcrumbBarItemClickedEventArgs args)
+    {
+        if (args.Index < Breadcrumbs.Count - 1)
+        {
+            var crumb = (Breadcrumb)args.Item;
+            crumb.NavigateTo();
+        }
     }
 
     [RelayCommand]
