+++ settings/DevHome.Settings/ViewModels/SettingsViewModel.cs
@@ -3,15 +3,19 @@
 
 using System.Collections.ObjectModel;
 using CommunityToolkit.Mvvm.ComponentModel;
+using CommunityToolkit.Mvvm.Input;
 using DevHome.Common.Extensions;
 using DevHome.Common.Services;
 using DevHome.Settings.Models;
 using Microsoft.UI.Xaml;
+using Microsoft.UI.Xaml.Controls;
 
 namespace DevHome.Settings.ViewModels;
 
 public partial class SettingsViewModel : ObservableObject
 {
+    public ObservableCollection<Breadcrumb> Breadcrumbs { get; }
+
     [ObservableProperty]
     private ObservableCollection<SettingViewModel> _settingsList = new();
 
@@ -31,6 +35,21 @@ public partial class SettingsViewModel : ObservableObject
         foreach (var setting in settings)
         {
             SettingsList.Add(new SettingViewModel(setting, this));
+        }
+
+        Breadcrumbs = new ObservableCollection<Breadcrumb>
+        {
+            new(stringResource.GetLocalized("Settings_Header"), typeof(SettingsViewModel).FullName!),
+        };
+    }
+
+    [RelayCommand]
+    public void BreadcrumbBarItemClicked(BreadcrumbBarItemClickedEventArgs args)
+    {
+        if (args.Index < Breadcrumbs.Count - 1)
+        {
+            var crumb = (Breadcrumb)args.Item;
+            crumb.NavigateTo();
         }
     }
