+++ settings/DevHome.Settings/Views/AccountsPage.xaml
@@ -2,15 +2,15 @@
     x:Class="DevHome.Settings.Views.AccountsPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:converters="using:CommunityToolkit.WinUI.Converters"
-    xmlns:viewmodels="using:DevHome.Settings.ViewModels"
-    xmlns:models="using:DevHome.Settings.Models"
     xmlns:ctControls="using:CommunityToolkit.WinUI.Controls"
+    xmlns:i="using:Microsoft.Xaml.Interactivity"
+    xmlns:ic="using:Microsoft.Xaml.Interactions.Core"
+    xmlns:models="using:DevHome.Settings.Models"
+    xmlns:viewmodels="using:DevHome.Settings.ViewModels"
     xmlns:behaviors="using:DevHome.Common.Behaviors"
-    behaviors:NavigationViewHeaderBehavior.HeaderMode="Never"
-    mc:Ignorable="d">
+    behaviors:NavigationViewHeaderBehavior.HeaderContext="{x:Bind ViewModel}">
+
     <Page.Resources>
         <converters:DoubleToVisibilityConverter x:Key="CountToVisibilityConverter" GreaterThan="0" FalseValue="Collapsed" TrueValue="Visible"/>
 
@@ -35,21 +35,26 @@
         </DataTemplate>
     </Page.Resources>
 
+    <behaviors:NavigationViewHeaderBehavior.HeaderTemplate>
+        <DataTemplate x:DataType="viewmodels:AccountsViewModel">
+            <Grid MaxWidth="{ThemeResource MaxPageContentWidth}" Margin="{ThemeResource ContentPageMargin}">
+                <BreadcrumbBar
+                    x:Name="BreadcrumbBar"
+                    ItemsSource="{x:Bind Breadcrumbs}">
+                    <i:Interaction.Behaviors>
+                        <ic:EventTriggerBehavior EventName="ItemClicked">
+                            <ic:InvokeCommandAction Command="{x:Bind BreadcrumbBarItemClickedCommand}" />
+                        </ic:EventTriggerBehavior>
+                    </i:Interaction.Behaviors>
+                </BreadcrumbBar>
+            </Grid>
+        </DataTemplate>
+    </behaviors:NavigationViewHeaderBehavior.HeaderTemplate>
+
     <Grid
         MaxWidth="{ThemeResource MaxPageContentWidth}"
         Margin="{ThemeResource ContentPageMargin}">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="*" />
-        </Grid.RowDefinitions>
-
-        <BreadcrumbBar
-            x:Name="BreadcrumbBar"
-            Margin="0,0,0,16"
-            ItemClicked="BreadcrumbBar_ItemClicked"
-            ItemsSource="{x:Bind Breadcrumbs}" />
-
-        <ScrollViewer Grid.Row="1" VerticalAlignment="Top">
+        <ScrollViewer VerticalAlignment="Top">
             <StackPanel>
                 <ctControls:SettingsCard x:Uid="Settings_Accounts_AddAccount">
                     <ctControls:SettingsCard.HeaderIcon>
