+++ settings/DevHome.Settings/Views/ExperimentalFeaturesPage.xaml
@@ -3,36 +3,48 @@
     x:Class="DevHome.Settings.Views.ExperimentalFeaturesPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:behaviors="using:DevHome.Common.Behaviors"
-    xmlns:controls="using:CommunityToolkit.WinUI.Controls"
     xmlns:converters="using:CommunityToolkit.WinUI.Converters"
-    xmlns:models="using:DevHome.Common.Models"
-    xmlns:ic="using:Microsoft.Xaml.Interactions.Core"
+    xmlns:ctControls="using:CommunityToolkit.WinUI.Controls"
     xmlns:i="using:Microsoft.Xaml.Interactivity"
-    behaviors:NavigationViewHeaderBehavior.HeaderMode="Never">
+    xmlns:ic="using:Microsoft.Xaml.Interactions.Core"
+    xmlns:models="using:DevHome.Common.Models"
+    xmlns:viewmodels="using:DevHome.Settings.ViewModels"
+    xmlns:behaviors="using:DevHome.Common.Behaviors"
+    behaviors:NavigationViewHeaderBehavior.HeaderContext="{x:Bind ViewModel}">
+
     <Page.Resources>
-        <converters:DoubleToVisibilityConverter x:Key="CountToVisibilityConverter" GreaterThan="0" FalseValue="Visible" TrueValue="Collapsed"/>
+        <converters:DoubleToVisibilityConverter x:Key="CountToVisibilityConverter" GreaterThan="0" FalseValue="Visible" TrueValue="Collapsed" />
     </Page.Resources>
+
+    <behaviors:NavigationViewHeaderBehavior.HeaderTemplate>
+        <DataTemplate x:DataType="viewmodels:ExperimentalFeaturesViewModel">
+            <Grid MaxWidth="{ThemeResource MaxPageContentWidth}" Margin="{ThemeResource ContentPageMargin}">
+                <BreadcrumbBar
+                    x:Name="BreadcrumbBar"
+                    ItemsSource="{x:Bind Breadcrumbs}">
+                    <i:Interaction.Behaviors>
+                        <ic:EventTriggerBehavior EventName="ItemClicked">
+                            <ic:InvokeCommandAction Command="{x:Bind BreadcrumbBarItemClickedCommand}" />
+                        </ic:EventTriggerBehavior>
+                    </i:Interaction.Behaviors>
+                </BreadcrumbBar>
+            </Grid>
+        </DataTemplate>
+    </behaviors:NavigationViewHeaderBehavior.HeaderTemplate>
+
     <Grid
         MaxWidth="{ThemeResource MaxPageContentWidth}"
         Margin="{ThemeResource ContentPageMargin}">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="*" />
-        </Grid.RowDefinitions>
-
-        <BreadcrumbBar
-            x:Name="BreadcrumbBar"
-            Margin="0,0,0,16"
-            ItemClicked="BreadcrumbBar_ItemClicked"
-            ItemsSource="{x:Bind Breadcrumbs}" />
-
-        <ScrollViewer Grid.Row="1" VerticalAlignment="Top">
-            <StackPanel Orientation="Vertical">
+        <ScrollViewer VerticalAlignment="Top">
+            <StackPanel>
                 <ItemsRepeater ItemsSource="{x:Bind ViewModel.ExperimentalFeatures}">
                     <ItemsRepeater.ItemTemplate>
                         <DataTemplate x:DataType="models:ExperimentalFeature">
-                            <controls:SettingsCard Header="{x:Bind Name, Mode=OneWay}" Description="{x:Bind Description, Mode=OneWay}" Margin="{ThemeResource SettingsCardMargin}" Visibility="{x:Bind IsVisible, Mode=OneWay}">
+                            <ctControls:SettingsCard
+                                Header="{x:Bind Name, Mode=OneWay}"
+                                Description="{x:Bind Description, Mode=OneWay}"
+                                Margin="{ThemeResource SettingsCardMargin}"
+                                Visibility="{x:Bind IsVisible, Mode=OneWay}">
                                 <ToggleSwitch IsOn="{x:Bind IsEnabled, Mode=OneWay}" Grid.Column="1">
                                     <i:Interaction.Behaviors>
                                         <ic:EventTriggerBehavior EventName="Toggled">
@@ -40,11 +52,13 @@
                                         </ic:EventTriggerBehavior>
                                     </i:Interaction.Behaviors>
                                 </ToggleSwitch>
-                            </controls:SettingsCard>
+                            </ctControls:SettingsCard>
                         </DataTemplate>
                     </ItemsRepeater.ItemTemplate>
                 </ItemsRepeater>
-                <TextBlock x:Uid="Settings_ExperimentalFeatures_NoExperimentalFeatures" Visibility="{x:Bind ViewModel.ExperimentalFeatures.Count, Converter={StaticResource CountToVisibilityConverter}, Mode=OneWay}"/>
+                <TextBlock
+                    x:Uid="Settings_ExperimentalFeatures_NoExperimentalFeatures"
+                    Visibility="{x:Bind ViewModel.ExperimentalFeatures.Count, Converter={StaticResource CountToVisibilityConverter}, Mode=OneWay}" />
             </StackPanel>
         </ScrollViewer>
     </Grid>
