+++ settings/DevHome.Settings/Views/SettingsPage.xaml
@@ -5,32 +5,36 @@
     x:Class="DevHome.Settings.Views.SettingsPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:ctControls="using:CommunityToolkit.WinUI.Controls"
-    xmlns:settings="using:DevHome.Settings.ViewModels"
+    xmlns:i="using:Microsoft.Xaml.Interactivity"
+    xmlns:ic="using:Microsoft.Xaml.Interactions.Core"
+    xmlns:viewmodels="using:DevHome.Settings.ViewModels"
     xmlns:behaviors="using:DevHome.Common.Behaviors"
-    behaviors:NavigationViewHeaderBehavior.HeaderMode="Never"
-    mc:Ignorable="d">
+    behaviors:NavigationViewHeaderBehavior.HeaderContext="{x:Bind ViewModel}">
+
+    <behaviors:NavigationViewHeaderBehavior.HeaderTemplate>
+        <DataTemplate x:DataType="viewmodels:SettingsViewModel">
+            <Grid MaxWidth="{ThemeResource MaxPageContentWidth}" Margin="{ThemeResource ContentPageMargin}">
+                <BreadcrumbBar
+                    x:Name="BreadcrumbBar"
+                    ItemsSource="{x:Bind Breadcrumbs}">
+                    <i:Interaction.Behaviors>
+                        <ic:EventTriggerBehavior EventName="ItemClicked">
+                            <ic:InvokeCommandAction Command="{x:Bind BreadcrumbBarItemClickedCommand}" />
+                        </ic:EventTriggerBehavior>
+                    </i:Interaction.Behaviors>
+                </BreadcrumbBar>
+            </Grid>
+        </DataTemplate>
+    </behaviors:NavigationViewHeaderBehavior.HeaderTemplate>
 
     <Grid
         MaxWidth="{ThemeResource MaxPageContentWidth}"
         Margin="{ThemeResource ContentPageMargin}">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto" />
-            <RowDefinition Height="*" />
-        </Grid.RowDefinitions>
-
-        <BreadcrumbBar
-            x:Name="BreadcrumbBar"
-            Margin="0,0,0,16"
-            ItemClicked="BreadcrumbBar_ItemClicked"
-            ItemsSource="{x:Bind Breadcrumbs}" />
-
         <ScrollViewer Grid.Row="1" VerticalAlignment="Top">
             <ItemsRepeater ItemsSource="{x:Bind ViewModel.SettingsList, Mode=OneWay}">
                 <ItemsRepeater.ItemTemplate>
-                    <DataTemplate x:DataType="settings:SettingViewModel">
+                    <DataTemplate x:DataType="viewmodels:SettingViewModel">
                         <ctControls:SettingsCard Header="{x:Bind Header}" Description="{x:Bind Description}"
                                        IsClickEnabled="True" Command="{x:Bind NavigateSettingsCommand}" Margin="{ThemeResource SettingsCardMargin}">
                             <ctControls:SettingsCard.HeaderIcon>
